Made MultiPageNavigationBehavior generic so it can be extended and provide better support for CarouselPage which uses ContentPage and not a Page.
