+++ Source/Xamarin/Prism.Forms/Behaviors/CarouselPageNavigationBehavior.cs
@@ -1,0 +1,8 @@
+﻿using Xamarin.Forms;
+
+namespace Prism.Behaviors
+{
+    public class CarouselPageNavigationBehavior : MultiPageNavigationBehavior<ContentPage>
+    {
+    }
+}
